Move `Fetch` and `Submit` intermediate objects inside `Http`
