+++ lib/exoskeleton/src/core/exoskeleton-core.scala
@@ -86,11 +86,11 @@ package executives:
     type CliType = CliInvocation
 
     def invocation
-       (arguments:     Iterable[Text],
-        environment:     Environment,
+       (arguments:        Iterable[Text],
+        environment:      Environment,
         workingDirectory: WorkingDirectory,
-        stdio:         Stdio,
-        signals:       Spool[Signal])
+        stdio:            Stdio,
+        signals:          Spool[Signal])
        (using interpreter: CliInterpreter)
     :     CliInvocation =
