+++ lib/telekinesis/src/core/telekinesis-core.scala
@@ -40,10 +40,10 @@ import language.dynamics
 given Realm = realm"telekinesis"
 
 extension [EndpointType: Fetchable](endpoint: EndpointType)
-  def fetch: Fetch[EndpointType.Target] =
-    Fetch
+  def fetch: Http.Fetch[EndpointType.Target] =
+    Http.Fetch
      (EndpointType.text(endpoint), EndpointType.target(endpoint), EndpointType.hostname(endpoint))
 
-  def submit: Submit[EndpointType.Target] =
-    Submit
+  def submit: Http.Submit[EndpointType.Target] =
+    Http.Submit
      (EndpointType.text(endpoint), EndpointType.target(endpoint), EndpointType.hostname(endpoint))
